requirements.txt
keras
streamlit
pillow
numpy
pandas
requests
matplotlib
tensorflow
seaborn
plotly
streamlit